pyproject.toml
[project]
name = "pyxora"
version = "0.5.0.dev0"
authors = [{ name = "AntonisDevStuff", email = "antonisdevstuff@gmail.com" }]
description = "A game engine library using pygame-ce, pymunk, and more"
keywords = ["engine", "game engine", "game development"]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pygame-ce==2.5.5",
    "pymunk==7.1.0",
    "pdoc==15.0.3",
    "cx_Freeze==8.3.0",
    "pygbag==0.9.2",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries :: pygame",
]
[project.urls]
Homepage = "https://pyxora.github.io/pyxora-website"
Documentation = "https://pyxora.github.io/pyxora-docs"
Repository = "https://github.com/pyxora/pyxora-engine"
Issues = "https://github.com/pyxora/pyxora-engine/issues"
Releases = "https://github.com/pyxora/pyxora-engine/releases"

[project.scripts]
pyxora = "pyxora.__main__:main"

[tool.setuptools]
package-dir = { "pyxora" = "src" }
packages = [
    "pyxora",
    "pyxora.wrapper",
    "pyxora.utils",
    "pyxora.docs",
    "pyxora.data",
    "pyxora.examples",
    "pyxora.projects",
]

[tool.setuptools.package-data]
pyxora = ["data/**", "examples/*/**"]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"
